Possible new option parser for filters
In /misc/ because I don't know where to put it for now, will be moved at
the end if it gets used.

Inspired by perl's Getopt::Long's GetOptions, receives a linked list
of key/value pairs (preparsed, a true GetOptions clone would receive the
string and do the parsing itself) and a variadic list of option name and
reference to the variable where to store each option. Due to C
limitations, must be NULL terminated.